RavenDB-22181: We are checking too much for disposing of JsonOperationContext.
This has a huge cost. Latest profiling runs are showing that checking Dispose status collectively is the TOP 1 more costly operation on the Own Time metric, followed by `BlittableJsonDocumentBuilder.ReadInternal()` and `.GetLazyString()`. Most of those calls are redundant in a world view where operation contexts are not jumping threads. Therefore, we are moving those high frequency checks to the debug build and only do explicit runtime checks in async codepaths (as we are already doing).